iscsi_ibft, be2iscsi, iscsi_boot: fix boot kobj data lifetime management
Patchwork-id: 37683
O-Subject: [RHEL6.2 PATCH 20/61] iscsi_ibft, be2iscsi,
	iscsi_boot: fix boot kobj data lifetime management
Bugzilla: 647268 694098 696760 698288
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <michaelc@cs.wisc.edu>

be2iscsi passes the boot functions its phba object which is
allocated in the shost, but iscsi_ibft passes in a object
allocated for each item to display. The problem is that
iscsi_boot_sysfs was managing the lifetime of the object
passed in and doing a kfree on release. This causes a double
free for be2iscsi which frees the shost in its pci_remove.

This patch fixes the problem by adding a release callback
which the drivers can call kfree or a put() type of function
(needed for be2iscsi which will do a get/put on the shost).

Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>