vfs: Block mmapped writes while the fs is frozen
Patchwork-id: 37757
O-Subject: [RHEL6 PATCH 2/3] vfs: Block mmapped writes while the fs is frozen
Bugzilla: 692167
RH-Acked-by: Josef Bacik <josef@redhat.com>

Bug 692167 - [6.0] Deadlock between "flush" kthread and ioctl() for ext4.

Tested with reproducer in bug, and xfstests on 4k and 1k blocks

Backport of this commit:

From: Jan Kara <jack@suse.cz>
Date: Mon, 23 May 2011 22:23:35 +0000 (+0200)
Subject: vfs: Block mmapped writes while the fs is frozen
X-Git-Tag: v3.0-rc1~141^2~19
X-Git-Url: http://git.kernel.org/?p=linux%2Fkernel%2Fgit%2Ftorvalds%2Flinux-2.6.git;a=commitdiff_plain;h=ea13a86463fd0c26c2c209c53dc46b8eff81bad4

vfs: Block mmapped writes while the fs is frozen

We should not allow file modification via mmap while the filesystem is
frozen. So block in block_page_mkwrite() while the filesystem is frozen.
We cannot do the blocking wait in __block_page_mkwrite() since e.g. ext4
will want to call that function with transaction started in some cases
and that would deadlock. But we can at least do the non-blocking reliable
check in __block_page_mkwrite() which is the hardest part anyway.

We have to check for frozen filesystem with the page marked dirty and under
page lock with which we then return from ->page_mkwrite(). Only that way we
cannot race with writeback done by freezing code - either we mark the page
dirty after the writeback has started, see freezing in progress and block, or
writeback will wait for our page lock which is released only when the fault is
done and then writeback will writeout and writeprotect the page again.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>