GFS2: Print warning rather than calling assert during direct I/O page invalidation race
Patchwork-id: 38240
O-Subject: [RHEL 6.2] GFS2: Print warning rather than calling assert during direct I/O page invalidation race (bz #697019)
Bugzilla: 697019
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

When there is a race between direct I/O and buffered I/O this can result
in a race during page invalidation. This is generally harmless and never
likely to occur since mixing direct I/O and buffered I/O on the same
file at the same time is a rather strange thing to want to do.

A better fix for this would be to add a ->launder_page function to the
GFS2 aops, but that is a much more tricky thing to do. This patch fixes
most of the problem with minimal change.

This patch is upstream (and has been for a while now) and has also been
tested on a previous RHEL6 test kernel.

This fixes bz #679019

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>