bnx2fc: Introduce interface structure for each vlan interface
Patchwork-id: 38243
O-Subject: [RHEL6.2 PATCH 03/14] bnx2fc: Introduce interface structure for each vlan interface
Bugzilla: 696773
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Bhanu Prakash Gollapudi <bprakash@broadcom.com>

Submitted to scsi-misc, yet to be accepted.
http://marc.info/?l=linux-scsi&m=130924311329943&w=2

Currently, bnx2fc has a hba structure that can work with only a single vlan
interface.  When there is a change in vlan id, it does not have the capability
to switch to different vlan interface. To solve this problem, a new structure
called 'interface' has been introduced, and each hba can now have multiple
interfaces, one per vlan id.

Most of the patch is a moving the interface specific fields from hba to the
interface structure, and appropriately modifying the dereferences. A list of
interfaces (if_list) is maintained along with adapter list. During a create
call, the interface structure is allocated and added to if_list and deleted &
freed on a destroy call.  Link events are propogated to all interfaces
belonging to the hba.

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>