autofs4 - use helper for need mount check
Patchwork-id: 38319
O-Subject: [RHEL-6.2 PATCH 03/36] autofs4 - use helper for need mount check
Bugzilla: 704941
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Ian Kent <raven@themaw.net>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=704941

This patch is a dependency of the vfs-automount patch series.
It is patch 3 of 9 cleanup patches meant to bring the autofs source
in line with the upstream source (as much as possible) to minimize
the changes needed to back port the vfs-automount patch series that
resolves this problem.

Define simple helper function for checking if we need to trigger
a mount.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>