autofs4 - rename unhashed to active in autofs4_lookup()
Patchwork-id: 38321
O-Subject: [RHEL-6.2 PATCH 05/36] autofs4 - rename unhashed to active in autofs4_lookup()
Bugzilla: 704941
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Ian Kent <raven@themaw.net>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=704941

This patch is a dependency of the vfs-automount patch series.
It is patch 5 of 9 cleanup patches meant to bring the autofs source
in line with the upstream source (as much as possible) to minimize
the changes needed to back port the vfs-automount patch series that
resolves this problem.

Rename the variable unhashed to active in autofs4_lookup() to
better reflect its usage.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>