autofs4 - cleanup active and expire lookup
Patchwork-id: 38322
O-Subject: [RHEL-6.2 PATCH 06/36] autofs4 - cleanup active and expire lookup
Bugzilla: 704941
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Ian Kent <raven@themaw.net>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=704941

This patch is a dependency of the vfs-automount patch series.
It is patch 6 of 9 cleanup patches meant to bring the autofs source
in line with the upstream source (as much as possible) to minimize
the changes needed to back port the vfs-automount patch series that
resolves this problem.

The lookup functions for active and expiring dentrys use parameters
that can be easily obtained on entry so we change the call to to
take just the dentry.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>