autofs4 - eliminate d_unhashed in path walk checks
Patchwork-id: 38323
O-Subject: [RHEL-6.2 PATCH 07/36] autofs4 - eliminate d_unhashed in path walk checks
Bugzilla: 704941
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Ian Kent <raven@themaw.net>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=704941

This patch is a dependency of the vfs-automount patch series.
It is patch 7 of 9 cleanup patches meant to bring the autofs source
in line with the upstream source (as much as possible) to minimize
the changes needed to back port the vfs-automount patch series that
resolves this problem.

The d_unhashed() function is used by __simple_empty() in the path walking
callbacks but autofs mount point dentrys should have no directories at
all so a list_empty() on d_subdirs should be (and is) sufficient. This
patch was originally posted because, in a later patch that was never
merged, the dentry was unhashed in ->d_revalidate() in order to send
mount requests to ->lookup(). But then we could not rely on d_unhased()
to give reliable results because it could be called at any time by any
code path.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>