autofs4 - rename dentry to active in autofs4_lookup_active()
Patchwork-id: 38324
O-Subject: [RHEL-6.2 PATCH 08/36] autofs4 - rename dentry to active in autofs4_lookup_active()
Bugzilla: 704941
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Ian Kent <raven@themaw.net>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=704941

This patch is a dependency of the vfs-automount patch series.
It is patch 8 of 9 cleanup patches meant to bring the autofs source
in line with the upstream source (as much as possible) to minimize
the changes needed to back port the vfs-automount patch series that
resolves this problem.

In autofs4_lookup_active() a declaration within the list traversal
loop uses a declaration that has the same name as the function
parameter.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>