autofs4 - reinstate last used update on access
Patchwork-id: 38348
O-Subject: [RHEL-6.2 PATCH 32/36] autofs4 - reinstate last used update on access
Bugzilla: 704941
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Ian Kent <raven@themaw.net>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=704941

When direct (and offset) mounts were introduced the the last used
timeout could no longer be updated in ->d_revalidate(). This is
because covered direct mounts would be followed over without calling
the autofs file system. As a result the definition of the busyness
check for all entries was changed to be "actually busy" being an open
file or working directory within the automount. But now we have a call
back in the follow so the last used update on any access can be
re-instated. This requires DCACHE_MANAGE_TRANSIT to always be set.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>