first time swap use results in heavy swapping
Patchwork-id: 38381
O-Subject: [RHEL6.2 PATCH 1/2] [s390x]kernel: first time swap use results in
	heavy swapping
Bugzilla: 722461

Description
-----------
System temporarily hangs for a period of time while a lot of
I/O is done to the swap device

System z is different in respect to the referenced bits
compared to other architectures. On e.g. x86 a page without
an active mapper is not referenced by definition, on System z
the referenced bit in the storage key can indicate that the
page is referenced even without an active mapper. The first
time a system with large amounts of anonymous memory runs
into memory shortage all pages on the active / inactive
lists will be found referenced on System z while only the
mapped pages appear to be referenced on x86. This causes the
memory management on System z to do a full sweep over all
page cache pages and start writeback for all of them.

Follow the x86 model and only check the referenced bit in
the storage key if the page has active mappers. That will
cause some of the pages in the page cache to appear to be
non-referenced and the page scan will stop in time.

Bugzilla
--------
BZ 722461
https://bugzilla.redhat.com/show_bug.cgi?id=722461

Upstream status of the patch
----------------------------
The patch is currently queued for inclusion into the next kernel
release:
http://148.100.100.18/cgi-bin/gitweb.cgi?p=linux-2.6.git;a=shortlog;h=refs/heads/for-linus

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>