pstore: Allow the user to explicitly choose a backend
Patchwork-id: 38453
O-Subject: [RHEL6 PATCH 27/28] pstore: Allow the user to explicitly choose a
	backend
Bugzilla: 696135
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Matthew Garret <magarret@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=696135

https://lkml.org/lkml/2011/7/18/302
Author: Matthew Garrett <mjg@redhat.com>
Date:   Thu Jul 21 16:57:55 2011 -0400

    [PATCH V2 4/9] pstore: Allow the user to explicitly choose a backend

    pstore only allows one backend to be registered at present, but the
    system may provide several. Add a parameter to allow the user to choose
    which backend will be used rather than just relying on load order.

    Signed-off-by: Matthew Garrett <mjg@redhat.com>

Posted but not yet committed.  Upstream maintainer seems ok with it but publicly
and privately.  Should make its way upstream and helps us work around a disabling
pstore for RHEL-6.2 until it matures more upstream.

Small difference with upstream is we explicitly set the default to be 'none' to
force it not to register with ERST.