i915: Fix opregion notifications
Patchwork-id: 38538
O-Subject: [PATCH 3/6] [RHEL6.2 PATCH] i915: Fix opregion notifications
Bugzilla: 722957
RH-Acked-by: Adam Jackson <ajax@redhat.com>
RH-Acked-by: Ben Skeggs <bskeggs@redhat.com>

From: Matthew Garrett <mjg@redhat.com>

bz: 722957

opregion-based platforms will send ACPI video event 0x80 for a range of
notification types for legacy compatibility. This is interpreted as a
display switch event, which may not be appropriate in the circumstances.
When we receive such an event we should make sure that the platform is
genuinely requesting a display switch before passing that event through
to userspace.

[ajax: rhel6 port, original https://patchwork.kernel.org/patch/969472/ ]

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Tested-by: Adam Jackson <ajax@redhat.com>
Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>